[refactor]: Simpify MUI Next link
