+++ src/components/text/index.ts
@@ -1,9 +1,7 @@
 export { BulletPoints } from "./BulletPoints";
 export { IconLink } from "./IconLink";
-export type {
-  MuiNextLinkProps,
-  NextLinkComposedProps,
-} from "./NextLinkComposed";
-export { MuiNextLink, NextLinkComposed } from "./NextLinkComposed";
+export { MuiNextLink, type MuiNextLinkProps } from "./MuiNextLink";
+export type { NextLinkComposedProps } from "./NextLinkComposed";
+export { NextLinkComposed } from "./NextLinkComposed";
 export { SkillChips } from "./SkillChips";
 export { Title } from "./styled";
